# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unity Catalog Governance API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/context": {
      "get": {
        "tags": [
          "context"
        ],
        "summary": "Workspace, environment, mode, and configured resources",
        "operationId": "getContext",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated or identity mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "context"
        ],
        "summary": "Authenticated actor and default execution identity",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated or identity mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "tags": [
          "context"
        ],
        "summary": "Implementation coverage and runtime availability of every capability",
        "operationId": "listCapabilities",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated or identity mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Actor": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "kind": {
            "$ref": "#/components/schemas/ActorKind"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AppRole"
            },
            "type": "array",
            "title": "Roles"
          },
          "verified_by": {
            "type": "string",
            "enum": [
              "user_token",
              "cli_profile",
              "fixture"
            ],
            "title": "Verified By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display",
          "kind",
          "roles",
          "verified_by"
        ],
        "title": "Actor"
      },
      "ActorKind": {
        "type": "string",
        "enum": [
          "user",
          "service_principal"
        ],
        "title": "ActorKind"
      },
...
      "AppMode": {
        "type": "string",
        "enum": [
          "fixture",
          "connected_readonly",
          "connected"
        ],
        "title": "AppMode"
      },
      "AppRole": {
        "type": "string",
        "enum": [
          "viewer",
          "steward",
          "access_admin",
          "auditor",
          "platform_admin"
        ],
        "title": "AppRole"
      },
...
      "CapabilitiesResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "CapabilitiesResponse"
      },
      "Capability": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "status": {
            "$ref": "#/components/schemas/CapabilityStatusValue"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requires": {
            "items": {
              "$ref": "#/components/schemas/CapabilityRequirement"
            },
            "type": "array",
            "title": "Requires"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "domain",
          "status",
          "reason",
          "requires",
          "checked_at"
        ],
        "title": "Capability"
      },
      "CapabilityRequirement": {
        "type": "string",
        "enum": [
          "sql_warehouse",
          "durable_store",
          "account_client",
          "user_token_scope"
        ],
        "title": "CapabilityRequirement"
      },
      "CapabilityStatusValue": {
        "type": "string",
        "enum": [
          "available",
          "not_configured",
          "insufficient_permissions",
          "unsupported_in_environment",
          "not_implemented",
          "temporarily_unavailable",
          "unknown"
        ],
        "title": "CapabilityStatusValue"
      },
...
      "Completeness": {
        "type": "string",
        "enum": [
          "complete",
          "truncated",
          "partial_visibility",
          "unknown"
        ],
        "title": "Completeness"
      },
      "Context": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/AppMode"
          },
          "mode_label": {
            "type": "string",
            "enum": [
              "Demo — synthetic data",
              "Read-only",
              "Editing enabled"
            ],
            "title": "Mode Label"
          },
          "environment_label": {
            "type": "string",
            "title": "Environment Label"
          },
          "workspace_host": {
            "type": "string",
            "title": "Workspace Host"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "managed_catalogs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Managed Catalogs"
          },
          "support_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Contact"
          },
          "bootstrap_allowlist_active": {
            "type": "boolean",
            "title": "Bootstrap Allowlist Active"
          },
          "warehouse_configured": {
            "type": "boolean",
            "title": "Warehouse Configured"
          },
          "durable_store_configured": {
            "type": "boolean",
            "title": "Durable Store Configured"
          },
          "account_client_configured": {
            "type": "boolean",
            "title": "Account Client Configured"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "mode_label",
          "environment_label",
          "workspace_host",
          "workspace_id",
          "managed_catalogs",
          "support_contact",
          "bootstrap_allowlist_active",
          "warehouse_configured",
          "durable_store_configured",
          "account_client_configured"
        ],
        "title": "Context"
      },
      "ContextResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/Context"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "ContextResponse"
      },
      "DataSource": {
        "type": "string",
        "enum": [
          "unity_catalog_api",
          "account_api",
          "system_table",
          "durable_store",
          "application",
          "fixture"
        ],
        "title": "DataSource"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_FAILED",
          "UNAUTHENTICATED",
          "IDENTITY_MISMATCH",
          "FORBIDDEN_ROLE",
          "FORBIDDEN_SCOPE",
          "MODE_READ_ONLY",
          "INSUFFICIENT_PRIVILEGES",
          "SOD_VIOLATION",
          "NOT_FOUND",
          "PLAN_STALE",
          "PLAN_EXPIRED",
          "PLAN_TAMPERED",
          "PLAN_INVALIDATED",
          "DUPLICATE_SUBMISSION",
          "RATE_LIMITED",
          "INTERNAL_ERROR",
          "NOT_IMPLEMENTED",
          "UNSUPPORTED",
          "NOT_CONFIGURED",
          "UPSTREAM_UNAVAILABLE",
          "OUTCOME_UNKNOWN"
        ],
        "title": "ErrorCode"
      },
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": false,
            "title": "Success"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "next_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Steps"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorResponse"
      },
      "Executor": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "service_principal",
              "user",
              "account_service_principal"
            ],
            "title": "Kind"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "display",
          "reason"
        ],
        "title": "Executor"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError"
      },
...
      "Identity": {
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/Actor"
          },
          "executor": {
            "$ref": "#/components/schemas/Executor"
          }
        },
        "type": "object",
        "required": [
          "actor",
          "executor"
        ],
        "title": "Identity"
      },
      "IdentityResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/Identity"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "meta"
        ],
        "title": "IdentityResponse"
      },
...
      "Meta": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scope"
              },
              {
                "type": "null"
              }
            ]
          },
          "completeness": {
            "$ref": "#/components/schemas/Completeness"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "source",
          "observed_at",
          "completeness",
          "limitations",
          "correlation_id"
        ],
        "title": "Meta"
      },
...
      "Scope": {
        "properties": {
          "catalog": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalog"
          },
          "schema": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          }
        },
        "type": "object",
        "title": "Scope"
      },
...
    }
  }
}